egl, dri: Propagate context priority hint to driver->CreateContext
Jump through the layers of abstraction between egl and dri in order to
feed the context priority attribute through to the backend. This
requires us to read the value from the base _egl_context, convert it to
a DRI attribute, parse it again in the generic context creater before
passing it to the driver as a function parameter.

Issues:

 * How to pass back the priority of the create context (as it may be
modified by the driver) back to EGL?

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>